+++ src/main/java/com/unitalk/common/service/EmployeeService.java
@@ -1,9 +1,5 @@
 package com.unitalk.common.service;
 
-import com.unitalk.emp.model.dto.ProfessorListItem;
-
-import java.util.List;
-
 public interface EmployeeService {
-    List<ProfessorListItem> getProfessors();
+
 }
