+++ src/main/java/com/unitalk/emp/controller/ProfessorAssignmentController.java
@@ -1,16 +1,15 @@
 package com.unitalk.emp.controller;
 
-import com.unitalk.common.service.EmployeeService;
-import com.unitalk.common.service.StudentService;
 import com.unitalk.emp.model.dto.ProfessorAssignmentListItem;
 import com.unitalk.emp.model.dto.ProfessorAssignmentRequest;
 import com.unitalk.emp.model.dto.ProfessorListItem;
 import com.unitalk.emp.model.dto.StudentListItem;
 import com.unitalk.emp.service.ProfessorAssignmentService;
+import com.unitalk.emp.service.ProfessorListService;
+import com.unitalk.emp.service.StudentListService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.awt.print.Pageable;
 import java.util.List;
 
 @RestController
@@ -18,28 +17,28 @@ import java.util.List;
 public class ProfessorAssignmentController {
 
     private final ProfessorAssignmentService professorAssignmentService;
-    private final EmployeeService employeeService;
-    private final StudentService studentService;
+    private final ProfessorListService professorListService;
+    private final StudentListService studentListService;
 
     @Autowired
-    public ProfessorAssignmentController(ProfessorAssignmentService professorAssignmentService, EmployeeService employeeService, StudentService studentService) {
+    public ProfessorAssignmentController(ProfessorAssignmentService professorAssignmentService, ProfessorListService professorListService, StudentListService studentListService) {
         this.professorAssignmentService = professorAssignmentService;
-        this.employeeService = employeeService;
-        this.studentService = studentService;
+        this.professorListService = professorListService;
+        this.studentListService = studentListService;
     }
 
     //전체 교수목록 조회
     @GetMapping("/list/professors")
     public List<ProfessorListItem> getProfessors() {
         // "교수" 부서에 속하는 교직원들의 정보를 가져옵니다.
-        List<ProfessorListItem> professors = employeeService.getProfessors();
+        List<ProfessorListItem> professors = professorListService.getProfessors();
         return professors;
     }
 
     //전체 학생목록 조회
     @GetMapping("/list/students")
     public List<StudentListItem> getStudents() {
-        return studentService.getAllStudents();
+        return studentListService.getAllStudents();
     }
 
     //전체 지도교수 배정 이력 조회
