+++ src/main/java/com/unitalk/emp/repository/StudentListRepository.java
@@ -1,10 +1,10 @@
-package com.unitalk.common.repository;
+package com.unitalk.emp.repository;
 
 import com.unitalk.common.model.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Long> {
+public interface StudentListRepository extends JpaRepository<Student, Long> {
 
 }
