+++ src/main/java/com/unitalk/emp/service/ProfessorListService.java
@@ -1,0 +1,11 @@
+package com.unitalk.emp.service;
+
+import com.unitalk.emp.model.dto.ProfessorListItem;
+
+import java.util.List;
+
+public interface ProfessorListService {
+    //교수목록 조회
+    List<ProfessorListItem> getProfessors();
+    
+}
