+++ src/main/java/com/unitalk/emp/service/StudentListService.java
@@ -1,0 +1,9 @@
+package com.unitalk.emp.service;
+
+import com.unitalk.emp.model.dto.StudentListItem;
+
+import java.util.List;
+
+public interface StudentListService {
+    List<StudentListItem> getAllStudents();
+}
